requirements.txt
streamlit
pandas
holidays
plotly
google-api-python-client
google-auth
google-auth-httplib2